factor out SExpr/SExprWriter classes out of smtlib backend, and also tidy them up/document them
